+++ libra-core/src/main/kotlin/xyz/ivyxjc/libra/core/models/UseCaseTxn.kt
@@ -1,12 +1,12 @@
 package xyz.ivyxjc.libra.core.models
 
 class UseCaseTxn : AbstractTransaction() {
-    var sourceId: Long = -1
-    var usecaseId: Long = -1
+    var sourceId: Int = -1
+    var usecaseId: Int = -1
     var usecaseName: String = ""
     var usecaseStatus: String = ""
     lateinit var gcGuid: String
-    var ucTxnId: Long = -1
+    var ucTxnId: Int = -1
     val attributes = mutableMapOf<String, String>()
     var trans: AbstractTransaction? = null
